Handling exception when category has subcategories.

+++ src/main/java/br/com/wikicode/serviceImpl/CategoryServiceImpl.java
@@ -6,16 +6,22 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import br.com.wikicode.domain.Category;
+import br.com.wikicode.domain.Subcategory;
+import br.com.wikicode.exception.IntegrityViolationException;
 import br.com.wikicode.exception.ObjectNotFoundException;
 import br.com.wikicode.exception.UnprocessableEntityException;
 import br.com.wikicode.reposiroty.CategoryRepository;
 import br.com.wikicode.service.CategoryService;
+import br.com.wikicode.service.SubcategoryService;
 
 @Service
 public class CategoryServiceImpl implements CategoryService {
 
 	@Autowired
 	private CategoryRepository categoryRepository; 
+	
+	@Autowired
+	private SubcategoryService subcategoryService;
 	
 	@Override
 	public List<Category> all() {
@@ -54,6 +60,11 @@ public class CategoryServiceImpl implements CategoryService {
 
 	@Override
 	public void delete(Integer id) {
+		
+		List<Subcategory> subcategories = subcategoryService.subcategoriesWhereHasCategoryId(id);
+		if (subcategories.size() != 0) {
+			throw new IntegrityViolationException("Essa categoria possui dependencias.");
+		}
 		Category category = find(id);
 		categoryRepository.delete(category);
 	}
